rearange test files
