+++ tests/test_visualization.py
@@ -1,13 +1,9 @@
 import pytest
-import sys
 import os
-import filecmp
 import matplotlib as mpl
 mpl.use('Agg')
 
 import trackintel as ti
-from trackintel.preprocessing import positionfixes
-from trackintel.preprocessing import staypoints
 
 
 class TestIO:
