+++ trackintel/visualization/__init__.py
@@ -1,0 +1,3 @@
+from .positionfixes import *
+from .staypoints import *
+from .triplegs import *
